# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "webhook-recovery",
    "summary": "Fair backlog burndown and consumer-defined replay policy after a provider outage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/simulation": {
      "post": {
        "summary": "Create Simulation",
        "description": "Create a simulation and seed its cast, in one transaction.\n\nSeeding belongs here because fan-out reads ``subscription``: a simulation\nwith no consumers accepts events and delivers them to nobody.\n\nCreating a simulation *is* Reset. Everything is namespaced by\n``simulation_id``, so runs persist side by side and one can be left running\nwhile another starts.",
        "operationId": "create_simulation_api_simulation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SimulationCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/{simulation_id}": {
      "get": {
        "summary": "Read Simulation",
        "operationId": "read_simulation_api_simulation__simulation_id__get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Simulation",
        "description": "Pause / resume / change speed / flip fair drain.\n\nPause, resume and speed change are all epoch rewrites -- one shared piece of\narithmetic in ``app.core.clock``, applied here and nowhere else. The order\nmatters: the speed change is applied to the *post-status* config, so\npausing and speeding up in one request behaves like doing them in sequence.",
        "operationId": "patch_simulation_api_simulation__simulation_id__patch",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/{simulation_id}/event": {
      "post": {
        "summary": "Create Event",
        "description": "Ledger one event and fan it out to every subscribed consumer.\n\nIn production the callers here are the provider's own internal services. In\nthe demo a background task in this same process plays them -- through\n``ingest_events`` directly rather than back through this route, which is the\nsame code path without a loopback socket. This route is the real interface,\nand that is the part of the claim that matters.",
        "operationId": "create_event_api_simulation__simulation_id__event_post",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/{simulation_id}/consumer": {
      "get": {
        "summary": "List Consumers",
        "description": "Consumers with their live counters, seeded at simulation creation.",
        "operationId": "list_consumers_api_simulation__simulation_id__consumer_get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsumerRead"
                  },
                  "title": "Response List Consumers Api Simulation  Simulation Id  Consumer Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/simulation/{simulation_id}/decisions": {
      "get": {
        "summary": "Read Decisions",
        "description": "The most recent terminal decisions, newest first.\n\nDeliberately not cursored: ``delivery.id`` is assigned at ingest, not at\ncompletion, so a ``since_id`` cursor over it would silently skip decisions.",
        "operationId": "read_decisions_api_simulation__simulation_id__decisions_get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/process/{process_id}/kill": {
      "post": {
        "summary": "Kill Process",
        "description": "Ask a process to die ungracefully, on its next heartbeat.\n\nNot a kill signal -- the API has no way to reach another container, and\nwould not want one. It raises a flag that the target reads back on the\nheartbeat it was already making, and the target chooses where to act on it\n(:meth:`app.core.runner.ProcessRunner.die`).\n\nSo the death is asynchronous, up to one heartbeat interval away, and\ngenuinely ungraceful: no drain, no lock release, no deregistration. A worker\nleaves a batch of leases behind for :mod:`app.conductor.reaper`; a leader\nconductor drops its advisory lock by dropping the connection under it.\n\n**There is no revive.** Both compose and Railway restart a process that\nexits non-zero, and the replacement registers under a fresh id -- so the\ndead row simply ages out of the liveness window on its own.",
        "operationId": "kill_process_api_process__process_id__kill_post",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Process Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventCreate": {
        "properties": {
          "event_type": {
            "type": "string",
            "maxLength": 64,
            "title": "Event Type"
          },
          "entity_key": {
            "type": "string",
            "maxLength": 128,
            "title": "Entity Key"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "entity_key"
        ],
        "title": "EventCreate",
        "description": "One event from the provider.\n\n``occurred_at`` is deliberately not a field: it is server-assigned from the\nsimulation's virtual clock. Letting a caller set it would let them backdate\nan event past a staleness bound, and staleness is a policy the *consumer*\nowns."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SimStatus": {
        "type": "string",
        "enum": [
          "running",
          "paused",
          "done"
        ],
        "title": "SimStatus"
      },
      "SimulationCreate": {
        "properties": {
          "scenario_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Name"
          },
          "speed_multiplier": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Multiplier"
          },
          "fair_drain_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fair Drain Enabled"
          },
          "global_attempts_per_s": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Attempts Per S"
          }
        },
        "type": "object",
        "title": "SimulationCreate"
      },
      "SimulationPatch": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "speed_multiplier": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Multiplier"
          },
          "fair_drain_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fair Drain Enabled"
          },
          "global_attempts_per_s": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Attempts Per S"
          },
          "outage_override": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outage Override",
            "description": "Force the outage on/off. Null in the response means the scenario script decides."
          }
        },
        "type": "object",
        "title": "SimulationPatch",
        "description": "Every field optional -- omitted means \"leave alone\".\n\n``status`` is how pause and resume are expressed: they are epoch rewrites,\nnot a separate endpoint."
      },
      "SimulationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "status": {
            "$ref": "#/components/schemas/SimStatus"
          },
          "speed_multiplier": {
            "type": "number",
            "title": "Speed Multiplier"
          },
          "fair_drain_enabled": {
            "type": "boolean",
            "title": "Fair Drain Enabled"
          },
          "global_attempts_per_s": {
            "type": "number",
            "title": "Global Attempts Per S"
          },
          "outage_override": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outage Override"
          },
          "virtual_now": {
            "type": "string",
            "format": "date-time",
            "title": "Virtual Now"
          },
          "virtual_now_s": {
            "type": "number",
            "title": "Virtual Now S"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "created_at_wall": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Wall"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scenario_name",
          "status",
          "speed_multiplier",
          "fair_drain_enabled",
          "global_attempts_per_s",
          "outage_override",
          "virtual_now",
          "virtual_now_s",
          "phase",
          "created_at_wall"
        ],
        "title": "SimulationRead"
      },
...
    }
  }
}